Add functionality to navigate between movie cards and home
All of the cards are clickable and will take the user to the movie chosen. The
'Home' link is also clickable and will navigate back to the main page. A styling
issue arose that I need to figure out that knocked the movie card out of the css
styling. Considered making the title itself only a link, but feel that the entire
card being clickable is more user friendly.